refactor: extract gameReducer
